requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
websocket-client==0.59.0
